requirements.txt
# packages in environment at /opt/anaconda3/envs/wrf-env:
#
# Name                    Version                   Build  Channel
basemap                   1.3.0            py38hc4b8d57_0    anaconda
basemap-data-hires        1.2.2                         0    conda-forge
blas                      1.0                         mkl    anaconda
bzip2                     1.0.8                h1de35cc_0  
ca-certificates           2020.11.8            h033912b_0    conda-forge
certifi                   2020.11.8        py38h50d1736_0    conda-forge
cftime                    1.2.1            py38hf1fa96c_0  
chardet                   3.0.4                    pypi_0    pypi
curl                      7.71.1               hb0a8c7a_1  
cycler                    0.10.0                   py38_0    anaconda
ffmpeg                    4.2.2                h97e5cf8_0  
freetype                  2.10.2               ha233b18_0    anaconda
geos                      3.8.0                hb1e8313_0    anaconda
gettext                   0.19.8.1             hb0f4f8b_2  
gmp                       6.1.2                hb37e062_1  
gmplot                    1.4.1                    pypi_0    pypi
gnutls                    3.6.5             h91ad68e_1002  
hdf4                      4.2.13               h39711bb_2  
hdf5                      1.10.4               hfa1e0ec_0  
idna                      2.10                     pypi_0    pypi
intel-openmp              2020.1                      216    anaconda
joblib                    0.17.0                     py_0  
jpeg                      9b                   he5867d9_2  
kiwisolver                1.2.0            py38h04f5b5a_0    anaconda
krb5                      1.18.2               h75d18d8_0  
lame                      3.100                h1de35cc_0  
lcms2                     2.11                 h92f6f08_0  
libcurl                   7.71.1               h8a08a2b_1  
libcxx                    10.0.0                        1  
libedit                   3.1.20191231         h1de35cc_1  
libffi                    3.3                  hb1e8313_2  
libgfortran               3.0.1                h93005f0_2  
libiconv                  1.16                 h1de35cc_0  
libnetcdf                 4.7.3                he3b6227_0  
libopus                   1.3.1                h1de35cc_0  
libpng                    1.6.37               ha441bb4_0    anaconda
libssh2                   1.9.0                ha12b0ac_1  
libtiff                   4.1.0                hcb84e12_1  
libvpx                    1.7.0                h378b8a2_0  
llvm-openmp               10.0.0               h28b9765_0  
lz4-c                     1.9.2                h79c402e_3  
matplotlib-base           3.2.2            py38h5670ca0_0    anaconda
mkl                       2019.4                      233    anaconda
mkl-service               2.3.0            py38hfbe908c_0    anaconda
mkl_fft                   1.1.0            py38hc64f4ea_0    anaconda
mkl_random                1.1.1            py38h959d312_0    anaconda
ncurses                   6.2                  h0a44026_1  
netcdf4                   1.5.3            py38h28545a2_0  
nettle                    3.4.1                h3018a27_0  
numpy                     1.19.1           py38h3b9f5b6_0    anaconda
numpy-base                1.19.1           py38hcfb5961_0    anaconda
olefile                   0.46                       py_0  
openh264                  2.1.0                hd9629dc_0  
openssl                   1.1.1h               haf1e3a3_0    conda-forge
pandas                    1.1.3            py38hb1e8313_0  
pillow                    7.2.0            py38ha54b6ba_0  
pip                       20.2.1                   py38_0  
proj                      6.2.1                hfd5b9e3_0    anaconda
pyparsing                 2.4.7                      py_0    anaconda
pyproj                    2.6.1.post1      py38h6d9c95c_1    anaconda
pyshp                     2.1.0                      py_0    anaconda
python                    3.8.5                h26836e1_0  
python-dateutil           2.8.1                      py_0    anaconda
python_abi                3.8                      1_cp38    conda-forge
pytz                      2020.1                     py_0  
readline                  8.0                  h1de35cc_0  
requests                  2.25.0                   pypi_0    pypi
scikit-learn              0.23.2           py38h959d312_0  
scipy                     1.5.2            py38h2515648_0  
setuptools                49.2.1                   py38_0  
six                       1.15.0                     py_0    anaconda
sqlite                    3.32.3               hffcf06c_0  
threadpoolctl             2.1.0              pyh5ca1d4c_0  
tk                        8.6.10               hb0a8c7a_0  
tornado                   6.0.4            py38h1de35cc_1    anaconda
urllib3                   1.26.2                   pypi_0    pypi
wheel                     0.34.2                   py38_0  
x264                      1!157.20191217       h1de35cc_0  
xz                        5.2.5                h1de35cc_0  
zlib                      1.2.11               h1de35cc_3  
zstd                      1.4.5                h41d2c2f_0  
